# agricole-api/openapi.yaml
openapi: 3.0.3
info:
  title: Agricultural Federation API
  description: API pour la gestion des collectivités agricoles et de leurs membres - TD Final Prog3
  version: 1.0.0

paths:
  /members:
    post:
      summary: Créer de nouveaux membres
      operationId: createMembers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Member'
      responses:
        '201':
          description: Membres créés avec succès
        '400':
          description: Conditions de parrainage ou de paiement non remplies

  /collectivities:
    post:
      summary: Créer de nouvelles collectivités
      operationId: createCollectivities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Collectivity'
      responses:
        '201':
          description: Collectivités créées avec succès
        '400':
          description: Critères d'approbation ou nombre de membres insuffisants

  /collectivities/{id}/identify:
    patch:
      summary: Identifier une collectivité (Fonctionnalité J)
      description: Attribuer un nom et un numéro unique. Une fois défini, ces champs sont immuables.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - number
              properties:
                name:
                  type: string
                  example: "Coopérative d'Analamanga"
                number:
                  type: string
                  example: "COOP-2026-001"
      responses:
        '200':
          description: Collectivité identifiée avec succès
        '400':
          description: Nom/Numéro déjà utilisé ou modification interdite (immutabilité)
        '404':
          description: Collectivité introuvable

components:
  schemas:
    Member:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date
        gender:
          type: string
          enum: [MALE, FEMALE]
        registrationFeePaid:
          type: boolean
        membershipDuesPaid:
          type: boolean
        referees:
          type: array
          items:
            type: string
          description: IDs des parrains

    Collectivity:
      type: object
      properties:
        id:
          type: string
        location:
          type: string
        federationApproval:
          type: boolean
        members:
          type: array
          items:
            type: string
          description: IDs des membres (minimum 10)
        structure:
          $ref: '#/components/schemas/Structure'

    Structure:
      type: object
      properties:
        president:
          type: string
        vicePresident:
          type: string
        treasurer:
          type: string
        secretary:
          type: string