# agricole-api/src/main/java/mg/federation/agricole_api/entity/Collectivity.java
package mg.federation.agricole_api.entity;

import lombok.Data;
import jakarta.persistence.*;
import java.util.List;

@Entity @Data
public class Collectivity {
    @Id @GeneratedValue(strategy = GenerationType.UUID)
    private String id;
    private String location;
    @OneToOne(cascade = CascadeType.ALL)
    private CollectivityStructure structure;
    @ManyToMany
    private List<Member> members;
    @Column(unique = true)
    private String name;
    @Column(unique = true)
    private String number;
}